Add deep linking functionality
custom javascripts can be added to the bottom of a page, before the footer, by
including them in the YAML declarations at the top of a template (see
'useful-notes-for-makers-grads' post for an example).

A contents table of deep links can be generated by including the javascripts
and the contents-table.html in the template (also see above post for an
example).

The anchorjs.js file also generates link symbols next to headings.